Module chumsky::debug

Utilities for debugging parsers.

“He was staring at the instruments with the air of one who is trying to convert Fahrenheit to centigrade in his head while his house is burning down.”

Structs

ParserInfo

Information about a specific parser.

Silent

A silent debugger that emits no debugging messages nor collects any debugging data.

Verbose

A verbose debugger that emits debugging messages to the console.

Enums

ParseEvent

An event that occurred during parsing.

Traits

DebuggerDeprecated

A trait implemented by parser debuggers.